:sparkles: 카테고리 추가와 페이저블 기능 수정

+++ src/main/java/com/sparta/wuzuzu/domain/community_posts/repository/Community_CategoryRepository.java
@@ -4,7 +4,7 @@ import com.sparta.wuzuzu.domain.community_posts.entity.Community_Category;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Community_CategoryRepository extends JpaRepository<Community_Category, Integer> {
+public interface Community_CategoryRepository extends JpaRepository<Community_Category, Long> {
 
     Optional<Community_Category> findByNameEquals(String name);
 }
